Move metadata (*mdsec, *md) & fsentry into own files
