+++ metsrw/__init__.py
@@ -3,5 +3,7 @@ LOGGER = logging.getLogger(__name__)
 LOGGER.addHandler(logging.NullHandler())
 
 from exceptions import *
+from fsentry import *
+from metadata import *
 from mets import *
 from utils import *
