+++ metsrw/mets.py
@@ -1,481 +1,15 @@
-#!/usr/bin/python2 -OO
-
 from datetime import datetime
 import logging
 from lxml import etree
-import os
-from random import randint
 import sys
 
 # This package
 import exceptions
+import fsentry
+import metadata
 import utils
 
 LOGGER = logging.getLogger(__name__)
-
-
-class FSEntry(object):
-    """
-    A class representing a filesystem entry - either a file or a directory.
-
-    When passed to a METSWriter instance, the tree of FSEntry objects will
-    be used to construct the <fileSec> and <structMap> elements of a
-    METS document.
-
-    Unless otherwise specified, an FSEntry object is assumed to be a file;
-    pass the `type` value as 'Directory' to specify that the object is
-    instead a directory.
-
-    An FSEntry object must be instantiated with a path as the first
-    argument to the constructor, which represents its path on disk.
-
-    An FSEntry object which is a Directory may have one or more children,
-    representing files or directories contained within itself. Directory
-    trees are designed for top-to-bottom traversal. Files cannot have
-    children, and attempting to instantiate a file FSEntry object with
-    children will raise a ValueError.
-
-    Any FSEntry object may have one or more metadata entries associated
-    with it; these can take the form of either references to other XML
-    files on disk, which should be wrapped in MDRef objects, or
-    wrapped copies of those XML files, which should be wrapped in
-    MDWrap objects.
-
-    :param str path: Path to the file on disk, as a bytestring. This will
-        populate FLocat @xlink:href
-    :param str label: Label in the structMap. If not provided, will be populated
-        with the basename of path
-    :param str use: Use for the fileGrp.  Items with identical uses will be
-        grouped together.
-    :param str type: Type of FSEntry this is. This will appear in the structMap.
-    :param list children: List of :class:`FSEntry` that are direct children of
-        this element in the structMap.  Only allowed if type is 'Directory'
-    :param str file_uuid: UUID of this entry. Will be used to construct the
-        FILEID used in the fileSec and structMap, and GROUPID.  Only required if
-        type is 'Item'.
-    :param FSEntry derived_from: FSEntry that this FSEntry is derived_from. This is used to set the GROUPID in the fileSec.
-    :raises ValueError: if children passed when type is not 'Directory'
-    """
-    def __init__(self, path, label=None, use='original', type=u'Item', children=None, file_uuid=None, derived_from=None):
-        # path can validly be any encoding; if this value needs
-        # to be spliced later on, it's better to treat it as a
-        # bytestring than as actually being encoded text.
-        self.path = str(path)
-        if label is None:
-            label = os.path.basename(path)
-        self.label = label
-        self.use = use
-        self.type = unicode(type)
-        if children is None:
-            children = []
-        self.children = children
-        self.file_uuid = file_uuid
-        self.derived_from = derived_from
-        self.amdsecs = []
-        self.dmdsecs = []
-
-        if type != 'Directory' and children:
-            raise ValueError("Only directory objects can have children")
-
-    def _create_id(self, prefix):
-        return prefix + '_' + str(randint(1, 999999))
-
-    def file_id(self):
-        """ Returns the fptr FILEID if this is not a Directory. """
-        if self.type == 'Directory':
-            return None
-        if self.file_uuid is None:
-            raise exceptions.MetsError('No FILEID: File %s does not have file_uuid set' % self.path)
-        return utils.FILE_ID_PREFIX + self.file_uuid
-
-    def group_id(self):
-        """
-        Returns the GROUPID.
-
-        If derived_from is set, returns that group_id.
-        """
-        if self.derived_from is not None:
-            return self.derived_from.group_id()
-        if self.file_uuid is None:
-            raise exceptions.MetsError('No GROUPID: File %s does not have file_uuid set' % self.path)
-        return utils.GROUP_ID_PREFIX + self.file_uuid
-
-    def admids(self):
-        """ Returns a list of ADMIDs for this entry. """
-        return [a.id_string() for a in self.amdsecs]
-
-    def dmdids(self):
-        """ Returns a list of DMDIDs for this entry. """
-        return [d.id_string() for d in self.dmdsecs]
-
-    def _add_metadata_element(self, md, subsection, mdtype, mode='mdwrap', **kwargs):
-        """
-        :param md: Value to pass to the MDWrap/MDRef
-        :param str subsection: Metadata tag to create.  See :const:`SubSection.ALLOWED_SUBSECTIONS`
-        :param str mdtype: Value for mdWrap/mdRef @MDTYPE
-        :param str mode: 'mdwrap' or 'mdref'
-        :param str loctype: Required if mode is 'mdref'. LOCTYPE of a mdRef
-        :param str label: Optional. Label of a mdRef
-        :param str otherloctype: Optional. OTHERLOCTYPE of a mdRef.
-
-        """
-        # HELP how handle multiple amdSecs?
-        # When adding *MD which amdSec to add to?
-        if mode.lower() == 'mdwrap':
-            mdsec = MDWrap(md, mdtype)
-        elif mode.lower() == 'mdref':
-            loctype = kwargs.get('loctype')
-            label = kwargs.get('label')
-            otherloctype = kwargs.get('otherloctype')
-            mdsec = MDRef(md, mdtype, loctype, label, otherloctype)
-        subsection = SubSection(subsection, mdsec)
-        if subsection.subsection == 'dmdSec':
-            self.dmdsecs.append(subsection)
-        else:
-            try:
-                amdsec = self.amdsecs[0]
-            except IndexError:
-                amdsec = AMDSec()
-                self.amdsecs.append(amdsec)
-            amdsec.subsections.append(subsection)
-        return subsection
-
-    def add_techmd(self, md, mdtype, mode='mdwrap', **kwargs):
-        return self._add_metadata_element(md, 'techMD', mdtype, mode, **kwargs)
-
-    def add_digiprovmd(self, md, mdtype, mode='mdwrap', **kwargs):
-        return self._add_metadata_element(md, 'digiprovMD', mdtype, mode, **kwargs)
-
-    def add_rightsmd(self, md, mdtype, mode='mdwrap', **kwargs):
-        return self._add_metadata_element(md, 'rightsMD', mdtype, mode, **kwargs)
-
-    def add_dmdsec(self, md, mdtype, mode='mdwrap', **kwargs):
-        return self._add_metadata_element(md, 'dmdSec', mdtype, mode, **kwargs)
-
-    def add_premis_object(self, md, mode='mdwrap'):
-        # TODO add extra args and create PREMIS object here
-        return self.add_techmd(md, 'PREMIS:OBJECT', mode)
-
-    def add_premis_event(self, md, mode='mdwrap'):
-        # TODO add extra args and create PREMIS object here
-        return self.add_digiprovmd(md, 'PREMIS:EVENT', mode)
-
-    def add_premis_agent(self, md, mode='mdwrap'):
-        # TODO add extra args and create PREMIS object here
-        return self.add_digiprovmd(md, 'PREMIS:AGENT', mode)
-
-    def add_premis_rights(self, md, mode='mdwrap'):
-        # TODO add extra args and create PREMIS object here
-        return self.add_rightsmd(md, 'PREMIS:RIGHTS', mode)
-
-    def add_dublin_core(self, md, mode='mdwrap'):
-        # TODO add extra args and create DC object here
-        return self.add_dmdsec(md, 'DC', mode)
-
-    def add_child(self, child):
-        if self.type != 'Directory':
-            raise ValueError("Only directory objects can have children")
-        self.children.append(child)
-
-
-class SubSection(object):
-    """
-    An object representing a metadata subsection in a document.
-
-    This is usually created automatically and does not have to be instantiated directly.
-
-    :param str subsection: Tag name for the subsection to be created.  Should be
-        one of 'techMD', 'rightsMD', 'sourceMD' or 'digiprovMD' if contained in an
-        :class:`amdSec`, or 'dmdSec'.
-    :param contents: The MDWrap or MDRef contained in this subsection.
-    :type contents: :class:`MDWrap` or :class:`MDRef`
-    :param str section_id: ID of the section. If not provided, will be generated from subsection tag and a random number.
-    """
-    ALLOWED_SUBSECTIONS = ('techMD', 'rightsMD', 'sourceMD', 'digiprovMD', 'dmdSec')
-
-    def __init__(self, subsection, contents, section_id=None):
-        if subsection not in self.ALLOWED_SUBSECTIONS:
-            raise ValueError(
-                '%s must be one of %s' % (subsection, self.ALLOWED_SUBSECTIONS))
-        self.subsection = subsection
-        self.contents = contents
-        self._id = section_id
-        self.status = None
-        self.older = None
-        self.newer = None
-        self.created = None
-
-    def __lt__(self, other):
-        # Sort based on the subsection's order in ALLOWED_SUBSECTIONS
-        # techMDs < rightsMD < sourceMD < digiprovMD < dmdSec
-        return self.ALLOWED_SUBSECTIONS.index(self.subsection) < self.ALLOWED_SUBSECTIONS.index(other.subsection)
-
-    def id_string(self, force_generate=False):
-        """
-        Returns the ID string for this SubSection.
-
-        :param bool force_generate: If True, will generate a new ID from the subsection tag and a random number.
-        """
-        if force_generate or not self._id:
-            self._id = self.subsection + '_' + str(randint(1, 999999))
-        return self._id
-
-    def get_status(self):
-        """
-        Returns the STATUS when serializing.
-
-        Calculates based on the subsection type and if it's replacing anything.
-
-        :returns: None or the STATUS string.
-        """
-        if self.status:
-            return self.status
-        if self.subsection == 'dmdSec':
-            if self.older is None:
-                return 'original'
-            else:
-                return 'updated'
-        if self.subsection == 'rightsMD':
-            # TODO how to handle ones where newer has been deleted?
-            if self.newer == None:
-                return 'current'
-            else:
-                return 'superseded'
-        return None
-
-
-    def replace_with(self, new_subsection):
-        """
-        Replace this SubSection with new_subsection.
-
-        Replacing SubSection must be the same time.  That is, you can only
-        replace a dmdSec with another dmdSec, or a rightsMD with a rightsMD etc.
-
-        :param new_subsection: Updated version of this SubSection
-        :type new_subsection: :class:`SubSection`
-        """
-        if self.subsection != new_subsection.subsection:
-            raise exceptions.MetsError('Must replace a SubSection with one of the same type.')
-        # TODO convert this to a DB so have bidirectonal foreign keys??
-        self.newer = new_subsection
-        new_subsection.older = self
-        self.status = None
-
-    @classmethod
-    def parse(cls, root):
-        """
-        Create a new SubSection by parsing root.
-
-        :param root: Element or ElementTree to be parsed into an object.
-        :raises exceptions.ParseError: If root's tag is not in :const:`SubSection.ALLOWED_SUBSECTIONS`.
-        :raises exceptions.ParseError: If the first child of root is not mdRef or mdWrap.
-        """
-        subsection = root.tag.replace(utils.lxmlns('mets'), '', 1)
-        if subsection not in cls.ALLOWED_SUBSECTIONS:
-            raise exceptions.exceptions.ParseError('SubSection can only parse elements with tag in %s with METS namespace' % cls.ALLOWED_SUBSECTIONS)
-        section_id = root.get('ID')
-        created = root.get('CREATED')
-        status = root.get('STATUS')
-        child = root[0]
-        if child.tag == utils.lxmlns('mets') + 'mdWrap':
-            mdwrap = MDWrap.parse(child)
-            obj = cls(subsection, mdwrap, section_id)
-        elif child.tag == utils.lxmlns('mets') + 'mdRef':
-            mdref = MDRef.parse(child)
-            obj = cls(subsection, mdref, section_id)
-        else:
-            raise exceptions.exceptions.ParseError('Child of %s must be mdWrap or mdRef' % subsection)
-        obj.created = created
-        obj.status = status
-        return obj
-
-    def serialize(self, now):
-        created = self.created or now
-        el = etree.Element(utils.lxmlns('mets') + self.subsection, ID=self.id_string(), CREATED=created)
-        status = self.get_status()
-        if status:
-            el.set('STATUS', status)
-        if self.contents:
-            el.append(self.contents.serialize())
-        return el
-
-
-class MDRef(object):
-    """
-    An object representing an external XML document, typically associated
-    with an :class:`FSEntry` object.
-
-    :param str target: Path to the external document. MDRef does not validate
-        the existence of this target.
-    :param str mdtype: The string representing the mdtype of XML document being
-        enclosed. Examples include "PREMIS:OBJECT" and "PREMIS:EVENT".
-    :param str label: Optional LABEL for the mdRef element
-    :param str loctype: LOCTYPE of the mdRef.  Must be one of 'ARK', 'URN', 'URL', 'PURL', 'HANDLE', 'DOI' or 'OTHER'.
-    :param str otherloctype: OTHERLOCTYPE of the mdRef. Should be provided if loctype is OTHER.
-    """
-    VALID_LOCTYPE = ('ARK', 'URN', 'URL', 'PURL', 'HANDLE', 'DOI', 'OTHER')
-
-    def __init__(self, target, mdtype, loctype, label=None, otherloctype=None):
-        self.target = target
-        self.mdtype = mdtype
-        self.loctype = loctype
-        if loctype not in self.VALID_LOCTYPE:
-            raise ValueError('loctype must be one of {}'.format(', '.join(self.VALID_LOCTYPE)))
-        self.label = label
-        self.otherloctype = otherloctype
-
-    @classmethod
-    def parse(cls, root):
-        """
-        Create a new MDWrap by parsing root.
-
-        :param root: Element or ElementTree to be parsed into a MDWrap.
-        """
-        if root.tag != utils.lxmlns('mets') + 'mdRef':
-            raise exceptions.ParseError('MDRef can only parse mdRef elements with METS namespace.')
-        # Required attributes
-        mdtype = root.get('MDTYPE')
-        if not mdtype:
-            raise exceptions.ParseError('mdRef must have a MDTYPE')
-        target = root.get(utils.lxmlns('xlink') + 'href')
-        if not target:
-            raise exceptions.ParseError('mdRef must have an xlink:href.')
-        loctype = root.get('LOCTYPE')
-        if not loctype:
-            raise exceptions.ParseError('mdRef must have a LOCTYPE')
-        # Optional attributes
-        label = root.get('LABEL')
-        otherloctype = root.get('OTHERLOCTYPE')
-
-        return cls(target, mdtype, loctype, label, otherloctype)
-
-    def serialize(self):
-        # If the source document is a METS document, the XPTR attribute of
-        # this mdRef element should point to the IDs of each dmdSec element
-        # in that document.
-        XPTR = None
-        try:
-            target_doc = etree.parse(self.target)
-            dmdsecs = [item.get('ID') for item in
-                       target_doc.findall(utils.lxmlns('mets')+'dmdSec')]
-            XPTR = "xpointer(id(''))".format(' '.join(dmdsecs))
-        except Exception:
-            pass
-
-        attrib = {
-            'MDTYPE': self.mdtype,
-            'LOCTYPE': self.loctype,
-        }
-        if self.target:
-            attrib[utils.lxmlns('xlink')+'href'] = self.target
-        if self.otherloctype:
-            attrib['OTHERLOCTYPE'] = self.otherloctype
-        if XPTR:
-            attrib['XPTR'] = XPTR
-        if self.label:
-            attrib['LABEL'] = self.label
-        return etree.Element(utils.lxmlns('mets') + 'mdRef', attrib=attrib)
-
-
-class MDWrap(object):
-    """
-    An object representing an XML document enclosed in a METS document.
-    The entirety of the XML document will be included; to reference an
-    external document, use the :class:`MDRef` class.
-
-    :param str document: A string copy of the document, and will be parsed into
-        an ElementTree at the time of instantiation.
-    :param str mdtype: The MDTYPE of XML document being enclosed. Examples
-        include "PREMIS:OBJECT" and "PREMIS:EVENT".
-    """
-    def __init__(self, document, mdtype):
-        parser = etree.XMLParser(remove_blank_text=True)
-        if isinstance(document, basestring):
-            self.document = etree.fromstring(document, parser=parser)
-        elif isinstance(document, etree._Element):
-            self.document = document
-        self.mdtype = mdtype
-
-    @classmethod
-    def parse(cls, root):
-        """
-        Create a new MDWrap by parsing root.
-
-        :param root: Element or ElementTree to be parsed into a MDWrap.
-        :raises exceptions.ParseError: If mdWrap does not contain MDTYPE
-        :raises exceptions.ParseError: If mdWrap or xmlData contain multiple children
-        """
-        if root.tag != utils.lxmlns('mets') + 'mdWrap':
-            raise exceptions.ParseError('MDWrap can only parse mdWrap elements with METS namespace.')
-        mdtype = root.get('MDTYPE')
-        if not mdtype:
-            raise exceptions.ParseError('mdWrap must have a MDTYPE')
-        document = root.xpath('mets:xmlData/*', namespaces=utils.NAMESPACES)
-        if len(document) != 1:
-            raise exceptions.ParseError('mdWrap and xmlData can only have one child')
-        document = document[0]
-        return cls(document, mdtype)
-
-    def serialize(self):
-        el = etree.Element(utils.lxmlns('mets') + 'mdWrap', MDTYPE=self.mdtype)
-        xmldata = etree.SubElement(el, utils.lxmlns('mets') + 'xmlData')
-        xmldata.append(self.document)
-
-        return el
-
-
-class AMDSec(object):
-    """
-    An object representing a section of administrative metadata in a
-    document.
-
-    This is ordinarily created by METSWriter instances and does not
-    have to be instantiated directly.
-
-    :param str section_id: ID of the section. If not provided, will be generated from 'amdSec' and a random number.
-    :param list subsections: List of :class:`SubSection` that are part of this amdSec
-    """
-    tag = 'amdSec'
-
-    def __init__(self, section_id=None, subsections=None):
-        if subsections is None:
-            subsections = []
-        self.subsections = subsections
-        self._id = section_id
-
-    def id_string(self, force_generate=False):
-        """
-        Returns the ID string for the amdSec.
-
-        :param bool force_generate: If True, will generate a new ID from 'amdSec' and a random number.
-        """
-        # e.g., amdSec_1
-        if force_generate or not self._id:
-            self._id = self.tag + '_' + str(randint(1, 999999))
-        return self._id
-
-    @classmethod
-    def parse(cls, root):
-        """
-        Create a new AMDSec by parsing root.
-
-        :param root: Element or ElementTree to be parsed into an object.
-        """
-        if root.tag != utils.lxmlns('mets') + 'amdSec':
-            raise exceptions.ParseError('AMDSec can only parse amdSec elements with METS namespace.')
-        section_id = root.get('ID')
-        subsections = []
-        for child in root:
-            subsection = SubSection.parse(child)
-            subsections.append(subsection)
-        return cls(section_id, subsections)
-
-    def serialize(self, now):
-        el = etree.Element(utils.lxmlns('mets') + self.tag, ID=self.id_string())
-        self.subsections.sort()
-        for child in self.subsections:
-            el.append(child.serialize(now))
-        return el
 
 
 class METSWriter(object):
@@ -752,7 +286,7 @@ class METSWriter(object):
             children = self._parse_tree_structmap(tree, elem)
 
             # Create FSEntry
-            fsentry = FSEntry(path=path, label=label, use=use, type=entry_type, children=children, file_uuid=file_uuid, derived_from=derived_from)
+            fs_entry = fsentry.FSEntry(path=path, label=label, use=use, type=entry_type, children=children, file_uuid=file_uuid, derived_from=derived_from)
 
             # Add DMDSecs
             dmdids = elem.get('DMDID')
@@ -760,18 +294,18 @@ class METSWriter(object):
                 dmdids = dmdids.split()
                 for dmdid in dmdids:
                     dmdsec_elem = tree.find('mets:dmdSec[@ID="' + dmdid + '"]', namespaces=utils.NAMESPACES)
-                    dmdsec = SubSection.parse(dmdsec_elem)
-                    fsentry.dmdsecs.append(dmdsec)
+                    dmdsec = metadata.SubSection.parse(dmdsec_elem)
+                    fs_entry.dmdsecs.append(dmdsec)
 
             # Add AMDSecs
             if amdids:
                 amdids = amdids.split()
                 for amdid in amdids:
                     amdsec_elem = tree.find('mets:amdSec[@ID="' + amdid + '"]', namespaces=utils.NAMESPACES)
-                    amdsec = AMDSec.parse(amdsec_elem)
-                    fsentry.amdsecs.append(amdsec)
+                    amdsec = metadata.AMDSec.parse(amdsec_elem)
+                    fs_entry.amdsecs.append(amdsec)
 
-            siblings.append(fsentry)
+            siblings.append(fs_entry)
         return siblings
 
     def _parse_tree(self, tree=None):
